zdfg

+++ .docker/entrypoint.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
 
 #On error no such file entrypoint.sh, execute in terminal - dos2unix .docker\entrypoint.sh
-cp .env.example .env
-cp .env.testing.example .env.testing
+cd backend
+if [ ! -f ".env" ]; then
+  cp .env.example .env
+if [ ! -f ".env.testing" ]; then
+  cp .env.testing.example .env.testing
+
 composer install
 php artisan key:generate
 php artisan migrate
